Add classes to each aspect of product in view
Give each part of the product in the index view file a class such as
'product-name' so the feature test can assert again how many of
that class there are. This allows to be more specific about how many
items are rendered on the page and that each attribute for the item
is there.